refactor: put all files/forder to crud-for-interns folder for avoid confusion
