+++ app/pages/admin/crud-for-interns/interns/[id].vue
