+++ app/server/api/crud-for-interns/batches/index.get.ts
@@ -1,0 +1,7 @@
+import { getAllBatchesUseCase } from '~/server/use-case/crud-for-interns/batches/get-all-batches.use-case'
+import type { RequestContext } from '~/server/types/RequestContext'
+
+export default defineEventHandler(async (event) => {
+  const batches = await getAllBatchesUseCase(event.context as RequestContext)
+  return batches
+})
