+++ app/server/api/crud-for-interns/interns/[id].put.ts
@@ -1,0 +1,6 @@
+import { updateInternDetailsUseCase } from '~/server/use-case/crud-for-interns/interns/update-intern-details.use-case'
+
+export default defineEventHandler(async (event) => {
+  const body = await readBody(event)
+  return updateInternDetailsUseCase(body)
+})
