--- app/server/api/interns/[id].get.ts
@@ -1,6 +1,0 @@
-import { getInternDetailsUseCase } from '~/server/use-case/interns/get-intern-details.use-case'
-
-export default defineEventHandler(async (event) => {
-  const params = getRouterParams(event)
-  return getInternDetailsUseCase({ id: params.id })
-})
