--- app/server/api/interns/[id].put.ts
@@ -1,6 +1,0 @@
-import { updateInternDetailsUseCase } from '~/server/use-case/interns/update-intern-details.use-case'
-
-export default defineEventHandler(async (event) => {
-  const body = await readBody(event)
-  return updateInternDetailsUseCase(body)
-})
