--- app/server/api/school/index.get.ts
@@ -1,7 +1,0 @@
-import { getAllSchoolsUseCase } from '~/server/use-case/schools/get-all-schools.use-case'
-import type { RequestContext } from '~/server/types/RequestContext'
-
-export default defineEventHandler(async (event) => {
-  const schools = await getAllSchoolsUseCase(event.context as RequestContext)
-  return schools
-})
