+++ app/server/response/crud-for-interns/batches/batch-details.response.ts
