+++ app/server/response/crud-for-interns/batches/batch-list.response.ts
