+++ app/server/response/crud-for-interns/interns/intern-details.response.ts
