+++ app/server/response/crud-for-interns/schools/school-list.response.ts
