+++ app/server/service/crud-for-interns/interns/intern.service.ts
