+++ app/server/service/crud-for-interns/schools/school.service.ts
