+++ app/server/use-case/crud-for-interns/schools/get-all-schools.use-case.ts
@@ -1,0 +1,9 @@
+import { schoolService } from '~/server/service/crud-for-interns/schools/school.service'
+import type { RequestContext } from '~/server/types/RequestContext'
+import { schoolFactory } from '~/server/factory/crud-for-interns/schools/school.factory'
+
+export const getAllSchoolsUseCase = async (_context: RequestContext) => {
+  const rawSchools = await schoolService.getDistinctSchoolNames()
+  
+  return schoolFactory.toSchoolListResponse(rawSchools);
+}
