--- app/server/use-case/schools/get-all-schools.use-case.ts
@@ -1,9 +1,0 @@
-import { schoolService } from '~/server/service/schools/school.service'
-import type { RequestContext } from '~/server/types/RequestContext'
-import { schoolFactory } from '~/server/factory/schools/school.factory'
-
-export const getAllSchoolsUseCase = async (_context: RequestContext) => {
-  const rawSchools = await schoolService.getDistinctSchoolNames()
-  
-  return schoolFactory.toSchoolListResponse(rawSchools);
-}
