+++ app/types/crud-for-interns/Batch.ts
